requirements.txt
Django==1.11.8
Werkzeug==0.10.1
dj-database-url==0.3.0
django_cors_headers==2.2.0
django-extensions==1.9.6
django_filter==1.1.0
django_material==1.2.2
django-pipeline==1.6.13
django_viewflow==1.2.2
djangorestframework==3.7.7
enum34==1.0.4
futures==3.0.2
gevent==1.0.1
greenlet==0.4.5
gunicorn==19.2.1
Markdown==2.6.11
git+https://github.com/mysociety/multiple-django-popolo-sources.git@84ea209
-e git+https://github.com/public-people/django-popolo.git@filter-unknown-contact-types#egg=django-popolo
newrelic==2.46.0.37
pathlib==1.0.1
psycopg2==2.6
pyScss==1.3.4
six==1.9.0
social-auth-app-django==2.1.0
whitenoise==1.0.6
wsgiref==0.1.2
